פיצול מחלקת dal למחלקותת נפרדות למימוש של סוליד
